Bluetooth: hci: Fix wrong CTE request interval type
Request interval is a number of connection events that
is used to periodically run CTE request control procedure.
BT 5.3 Core Specification defines it as 2 octets long.
It had wrong type uint8_t. Changed to correct one uint16_t.

The commit also changes type of cte_rsp_en field of lll_df_conn_tx_cfg
to state that it is a boolean flag.

+++ include/bluetooth/direction.h
@@ -179,7 +179,7 @@ struct bt_df_conn_cte_req_params {
 	 * Value 0x0 means, run the procedure once. Other values are intervals in number of
 	 * connection events, to run the command periodically.
 	 */
-	uint8_t interval;
+	uint16_t interval;
 	/** Requested length of the CTE in 8 us units. */
 	uint8_t cte_length;
 	/**
